+++ src/index.js
@@ -4,6 +4,9 @@ const fs = require("fs-extra");
 const cloneDeepFunc = require("lodash.clonedeep");
 const JSZip = require("jszip");
 const RawSource = require("webpack-sources/lib/RawSource");
+const { parse } = require("@babel/parser");
+const traverse = require("@babel/traverse").default;
+const generate = require("@babel/generator").default;
 
 const { ENCODING } = require("./constants");
 const logger = require("./logger");
@@ -71,6 +74,7 @@ class PowerBICustomVisualsWebpackPlugin {
 			visualSourceLocation: "",
 			pluginLocation: path.join(".tmp", "precompile", "visualPlugin.ts"),
 			compression: 0, // no compression,
+			removeNetworkCalls: false,
 		};
 
 		this._name = "PowerBICustomVisualsWebpackPlugin";
@@ -187,12 +191,48 @@ class PowerBICustomVisualsWebpackPlugin {
 			const content = assets[asset].source();
 
 			if (extension === "js") {
-				assetsContent.jsContent = content;
+				assetsContent.jsContent = this.removeNetworkCalls(content, this.options.removeNetworkCalls)
 			} else if (extension === "css") {
 				assetsContent.cssContent = content;
 			}
 		}
 		return assetsContent;
+	}
+
+	removeNetworkCalls(code, shouldRemoveNetworkCalls = false) {
+        const parsedCode = parse(code, { sourceType: "module", plugins: ["jsx"] });
+		let numberOfRemovedNetworkCalls = 0;
+
+        traverse(parsedCode, {
+            CallExpression(path) {
+                const callee = path.get("callee");
+
+                if (callee.isIdentifier({ name: "fetch" })) {
+					if (shouldRemoveNetworkCalls) {
+						path.replaceWithSourceString("undefined");
+					}
+					numberOfRemovedNetworkCalls++;
+                }
+            },
+
+			NewExpression(path) {
+				const callee = path.get("callee");
+	
+				if (callee.isIdentifier({ name: "XMLHttpRequest" })) {
+					if (shouldRemoveNetworkCalls) {
+						path.replaceWithSourceString("undefined");
+					}
+					numberOfRemovedNetworkCalls++;
+				}
+			}
+        });
+
+		if (!shouldRemoveNetworkCalls && numberOfRemovedNetworkCalls > 0) {
+			logger.warn(`${numberOfRemovedNetworkCalls} possible network calls found. It is forbidden for visuals that are going to be PowerBI Certified.`);
+		} else if (numberOfRemovedNetworkCalls > 0) {
+			logger.warn(`${numberOfRemovedNetworkCalls} network calls were removed, test the visual before publishing`);
+		}
+        return generate(parsedCode, { retainLines: true }).code;
 	}
 
 	async _beforeCompile(callback) {